requirements.txt
alabaster==0.7.16
asn1crypto==1.5.1
asttokens==3.0.0
attr==0.3.2
babel==2.17.0
backcall==0.2.0
backports.tarfile==1.2.0
boltons==25.0.0
boto3==1.40.15
botocore==1.40.15
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
comm==0.2.3
configparser==7.2.0
contextlib2==21.6.0
cryptography==45.0.6
Cython==0.29.34
decorator==5.2.1
docutils==0.21
executing==2.2.0
filelock==3.19.1
funcsigs==1.0.2
HTMLParser==0.0.2
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
ipython==8.12.3
ipywidgets==8.1.7
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jnius==1.1.0
jupyterlab_widgets==3.0.15
keyring==25.6.0
lockfile==0.12.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mock==5.2.0
more-itertools==10.7.0
packaging==25.0
parso==0.8.4
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.32.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
railroad==0.5.0
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
snowballstemmer==3.0.1
snowflake-connector-python==3.17.2
sortedcontainers==2.4.0
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
thread==2.0.5
tomli==2.2.1
tomlkit==0.13.3
toolz==1.0.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
urllib3==2.5.0
urllib3-secure-extra==0.1.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
zipp==3.23.0
